refactor: extract shared constants and formatters
New modules:
- lib/constants/billing.ts — BILLING_STRATEGIES, BILLING_UNITS,
  PAYABLE_STATUSES, PAYMENT_METHODS, PAYMENT_METHOD_LABELS
- lib/constants/lease.ts — MOVE_OUT_REASONS, DEPOSIT_HANDLING_OPTIONS
- lib/formatters.ts additions — computeMonthlyEquivalent,
  formatPeriod, formatRupiah

Removed 260+ lines of inline duplicate definitions across
18 files. All files now import from the shared modules.
