+++ resources/js/components/features/leases/assign-unit-sheet.tsx
@@ -26,74 +26,13 @@ import {
     SheetHeader,
     SheetTitle,
 } from '@/components/ui/sheet';
+import { DUE_DAY_OPTIONS } from '@/lib/constants';
+import { BILLING_STRATEGIES } from '@/lib/constants/billing';
+import { computeMonthlyEquivalent, formatPrice } from '@/lib/formatters';
 import tenants from '@/routes/tenants';
 import type { AvailableUnit, UnitRate, Tenant } from '@/types';
 
 type AvailableUnits = AvailableUnit[];
-
-const BILLING_STRATEGIES = [
-    { value: 'advance', label: 'Advance (due within period)' },
-    { value: 'arrears', label: 'Arrears (due after period)' },
-];
-
-const DUE_DAY_OPTIONS = [
-    { value: '1', label: '1st' },
-    { value: '5', label: '5th' },
-    { value: '10', label: '10th' },
-    { value: '15', label: '15th' },
-    { value: '20', label: '20th' },
-    { value: '25', label: '25th' },
-    { value: '31', label: 'Last day of month' },
-];
-
-function formatCurrency(value: string | number): string {
-    const num = typeof value === 'string' ? Number.parseFloat(value) : value;
-
-    return new Intl.NumberFormat('id-ID', {
-        style: 'currency',
-        currency: 'IDR',
-        minimumFractionDigits: 0,
-        maximumFractionDigits: 0,
-    }).format(num);
-}
-
-function computeMonthlyEquivalent(
-    amount: string | null,
-    interval: number | null,
-    unit: string | null,
-): string {
-    if (!amount || !interval || !unit) {
-        return '';
-    }
-
-    const num = Number.parseFloat(amount);
-
-    if (isNaN(num)) {
-        return '';
-    }
-
-    const int = interval || 1;
-    let monthly: number;
-
-    switch (unit) {
-        case 'day':
-            monthly = (num * 365) / 12 / int;
-            break;
-        case 'week':
-            monthly = (num * 52) / 12 / int;
-            break;
-        case 'month':
-            monthly = num / int;
-            break;
-        case 'year':
-            monthly = num / 12 / int;
-            break;
-        default:
-            return '';
-    }
-
-    return `≈ ${formatCurrency(Math.round(monthly))}/month`;
-}
 
 export default function AssignUnitSheet({
     tenant,
@@ -373,7 +312,7 @@ export default function AssignUnitSheet({
                                                                 : ''}
                                                         </span>
                                                         <span className="font-medium tabular-nums">
-                                                            {formatCurrency(
+                                                            {formatPrice(
                                                                 rate.amount,
                                                             )}
                                                         </span>
