+++ resources/js/lib/constants/lease.ts
@@ -1,0 +1,13 @@
+export const MOVE_OUT_REASONS = [
+    { value: 'end_of_contract', label: 'End of Contract' },
+    { value: 'voluntary_move', label: 'Voluntary Move' },
+    { value: 'policy_violation', label: 'Policy Violation' },
+    { value: 'eviction', label: 'Eviction' },
+    { value: 'other', label: 'Other' },
+];
+
+export const DEPOSIT_HANDLING_OPTIONS = [
+    { value: 'refund', label: 'Refund deposit' },
+    { value: 'transfer', label: 'Transfer to new unit' },
+    { value: 'forfeit', label: 'Forfeit deposit' },
+];
