+++ resources/js/pages/dashboard/rent.tsx
@@ -7,6 +7,8 @@ import { SearchInput } from '@/components/data-table/search-input';
 import { Badge } from '@/components/ui/badge';
 import { Card, CardContent } from '@/components/ui/card';
 import { useTable } from '@/hooks/use-table';
+import { DUE_DAY_LABELS } from '@/lib/constants';
+import { formatPrice } from '@/lib/formatters';
 import { dashboard } from '@/routes';
 import { rent as dashboardRent } from '@/routes/dashboard';
 import type { PaginatedData, RentDashboardEntry, TableMeta } from '@/types';
@@ -28,33 +30,12 @@ type PageProps = {
     };
 };
 
-const DUE_DAY_LABELS: Record<number, string> = {
-    1: '1st',
-    5: '5th',
-    10: '10th',
-    15: '15th',
-    20: '20th',
-    25: '25th',
-    31: 'Last day',
-};
-
 const STATUS_BADGE: Record<string, { label: string; className: string }> = {
     paid: { label: 'Paid', className: 'bg-green-600 text-white' },
     overdue: { label: 'Overdue', className: 'bg-red-600 text-white' },
     due_today: { label: 'Due Today', className: 'bg-amber-600 text-white' },
     due_soon: { label: 'Due Soon', className: 'bg-blue-600 text-white' },
 };
-
-function formatPrice(cents: string): string {
-    const num = Number.parseFloat(cents);
-
-    return new Intl.NumberFormat('id-ID', {
-        style: 'currency',
-        currency: 'IDR',
-        minimumFractionDigits: 0,
-        maximumFractionDigits: 0,
-    }).format(num);
-}
 
 export default function Rent({
     entries: data,
