+++ resources/js/pages/properties/units/leases/index.tsx
@@ -3,6 +3,7 @@ import { useState } from 'react';
 import { LeaseDetailSheet, MoveOutSheet } from '@/components/features';
 import { Heading } from '@/components/shared';
 import { Badge } from '@/components/ui/badge';
+import { formatDate, formatPrice } from '@/lib/formatters';
 import properties from '@/routes/properties';
 import type { AvailableUnit, Lease, Property, Unit } from '@/types';
 
@@ -12,33 +13,6 @@ type PageProps = {
     leases: Lease[];
     availableUnits: AvailableUnit[];
 };
-
-function formatPrice(cents: string | null): string {
-    if (!cents) {
-        return '—';
-    }
-
-    const num = Number.parseFloat(cents);
-
-    return new Intl.NumberFormat('id-ID', {
-        style: 'currency',
-        currency: 'IDR',
-        minimumFractionDigits: 0,
-        maximumFractionDigits: 0,
-    }).format(num);
-}
-
-function formatDate(date: string | null): string {
-    if (!date) {
-        return '—';
-    }
-
-    return new Date(date).toLocaleDateString('id-ID', {
-        year: 'numeric',
-        month: 'short',
-        day: 'numeric',
-    });
-}
 
 const STATUS_COLORS: Record<string, string> = {
     active: 'bg-blue-600',
